UI: Adjustments for HQ buttons of join servers menu
Path to new assets. Also fixed the button size of the "specify", "update" and "delete" buttons (were bubble graphics in PadMod at that time). The "delete" button has been moved to the position of the old "create" button below the server list. This creates more space for the "Add to Favorites" and "Add All to Favorites" buttons, which are now positioned slightly lower. The order of the menu selection has been adapted to the new layout.